Refactoring: Rename and move picking utility script
